requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
coverage==7.2.7
exceptiongroup==1.2.2
fast-enum==1.3.0
fastapi==0.103.2
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.21.2
python-dotenv==0.21.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0

